requirements.txt
Django~=4.1.7
psycopg2-binary==2.9.1
djangorestframework~=3.14.0